Fix service worker errors (#337)
* Catch service worker promise rejection

When "Delete cookies and site data when Firefox is closed" setting is
enabled in FF the `navigator.serviceWorker.ready` promise rejects. When
unhandled this results in a large amount of error logs.

* Catch errors from postMessages to redundant service workers

This should not happen as `navigator.serviceWorker.ready` should only
resolve once service worker is in 'active' state. However, in Safari
there are cases where a `postMessage` even after the `ready` promise
will throw an InvalidState error indicating that the service worker is
in 'redundant' state.

* Update changelog

* Removed change log for change that was rolled back.

Co-authored-by: Mark Brocato <mark.brocato@moovweb.com>
Co-authored-by: Richard van der Dys <richard.vanderdys@gmail.com>